replace a well-intentioned &nbsp; with a ' ' instead (#1390)
The code looks to ensure there is a space between the end of the opengraph text and the expand/contract icon. However, after some refactorings, the &nbsp; in question ended up inside a string, and only conditionally displayed. Simplify all of this with just the necessary {' '} that accomplishes the desired purpose in React.