MM-12205: notify on session expiration (#1751)
* fix missing key on login controller

* dispatch logout vs. using Client4 directly

This allows the Redux store and associated listeners to respond to the
event vs. discovering it as a side effect of other requests failing.

* refactor notifications to allow for more options

* MM-12205: detect session expiration

When logging out, identify if the cause was a session expiration, and effectively redirect to /login?extra=expired, triggering controller behaviour there to display a notification (and tweak the document title).

* don't re-emit the notification once shown

* track focus/blur and (re-show) notification only when not focussed

This doesn't work 100% with the desktop app, where an initial load
doesn't trigger the onFocus. We can't assume it's already focussed,
though, since that could suppress a notification. Compromises.

* fix linting issues

* login --> log in

* once the notification has been shown, don't unshow it on blur

* track was logged via local storage instead

* fix focusing issues and notification dismissal

* preserve extra=expired interface with desktop app

* fix linting and unit tests

* tweak dismiss styling

* fix linting, snapshot errors